refactor(ws-do): split community machine frames
